RDKEMW-20406:[Flex 2.0][Xione Charter][E040.012.00.8.4p13s1][SYST_ERR_TLP999]Delay in moving the PlayerPipeline to 'Playing' state even after receiving the first video frame.

+++ AampEventManager.cpp
@@ -418,7 +418,7 @@ void AampEventManager::SendEventSync(const AAMPEventPtr &eventData)
 		SAFE_DELETE(pCurrent);
 	}
 #ifdef EVENT_DEBUGGING
-	AAMPLOG_WARN("[XSTLP-999-DBG][MAIN-LOOP-STARVATION] SendEventSync type=%d took %lld ms",eventType, (int)(NOW_STEADY_TS_MS - startTime));
+	AAMPLOG_WARN("[XSTLP-999-DBG][MAIN-LOOP-STARVATION] SendEventSync type=%d took %lld ms",eventType, (long long)(NOW_STEADY_TS_MS - startTime));
 #endif
 
 }
